Split unit tests from envtest tests
