+++ controllers/operator/suite_test.go
@@ -17,6 +17,7 @@ limitations under the License.
 package controllers
 
 import (
+	"os"
 	"path/filepath"
 	"testing"
 	"time"
@@ -47,6 +48,9 @@ var testEnv *envtest.Environment
 var nameGenerator namegenerator.Generator
 
 func TestAPIs(t *testing.T) {
+	if os.Getenv("RUN_ENVTEST") != "1" {
+		t.Skip("Skipping envtest tests")
+	}
 	RegisterFailHandler(Fail)
 
 	RunSpecsWithDefaultAndCustomReporters(t,
